requirements.txt
alpaca-py
alpaca_trade_api
numpy